MSVC 2008 Builds: Split Up the Property Sheets
Split up the property sheets, so that it is easier to maintain, and update
the project files accordingly.  Also clean up the project files by
finishing up the merge of the *_sdl.vcproj items into their regular
counterparts, and dropping all the *_sdl.* files.

Make use of UNIX line endings for all the MSVC 2008 build files, except the
.sln file and the README.txt files, for easier maintenance.